requirements.txt
numpy>=2.0.0
fastapi>=0.110.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.2.1
motor>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.9
typing-extensions>=4.9.0
pytest>=8.0.0
httpx>=0.26.0
langchain>=0.1.0
langchain-huggingface>=0.1.2
langchain-community>=0.0.13
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.36.0
google-generativeai>=0.3.1 